requirements.txt
aiohttp>=3.8.0
asyncssh>=2.13.2
PyYAML>=6.0
mysql-connector-python>=8.0.0
jinja2>=3.1.0
click>=8.1.0
rich>=13.0.0
aiofiles>=23.1.0 